BACKLOG-20460: Fixed permission check on listordering section
