+++ src/javascript/ContentEditor/editorTabs/EditPanelContent/FormBuilder/Sections/Section/Section.jsx
@@ -3,27 +3,10 @@ import {FieldSet} from '../../FieldSet';
 import React from 'react';
 import PropTypes from 'prop-types';
 import {SectionPropTypes} from '~/ContentEditor/ContentEditor.proptypes';
-
-const filterRegularFieldSets = fieldSets => {
-    const showFieldSet = fieldSet => {
-        if (!fieldSet || !fieldSet.visible) {
-            return false;
-        }
-
-        if (fieldSet.dynamic && !fieldSet.hasEnableSwitch && !fieldSet.activated) {
-            return false;
-        }
-
-        // We must hide fieldSet in the section when the fieldSet is not dynamic and
-        // the fieldSet doesn't contain any fields (empty).
-        return fieldSet.dynamic || fieldSet.fields.length > 0;
-    };
-
-    return fieldSets.filter(fs => showFieldSet(fs));
-};
+import {filterFieldSets} from '~/ContentEditor/editorTabs/EditPanelContent/FormBuilder/Sections';
 
 export const Section = ({section, isExpanded, onClick}) => {
-    const fieldSets = filterRegularFieldSets(section.fieldSets);
+    const fieldSets = filterFieldSets(section.fieldSets);
 
     if (fieldSets.length === 0) {
         return null;
