+++ src/javascript/ContentEditor/editorTabs/EditPanelContent/FormBuilder/Sections/filterFieldSets.jsx
@@ -1,0 +1,17 @@
+export const filterFieldSets = fieldSets => {
+    const showFieldSet = fieldSet => {
+        if (!fieldSet || !fieldSet.visible) {
+            return false;
+        }
+
+        if (fieldSet.dynamic && !fieldSet.hasEnableSwitch && !fieldSet.activated) {
+            return false;
+        }
+
+        // We must hide fieldSet in the section when the fieldSet is not dynamic and
+        // the fieldSet doesn't contain any fields (empty).
+        return fieldSet.dynamic || fieldSet.fields.length > 0;
+    };
+
+    return fieldSets.filter(fs => showFieldSet(fs));
+};
