+++ src/javascript/ContentEditor/editorTabs/EditPanelContent/FormBuilder/Sections/index.js
@@ -1,2 +1,3 @@
 export * from './Section';
 export * from './ChildrenSection';
+export * from './filterFieldSets';
